refactor: Improve code structure and reliability

+++ django_nitro_mailer/emails.py
@@ -1,42 +1,47 @@
 import logging
-from typing import Optional
+from dataclasses import dataclass
 
 from django.core.mail import get_connection
-from django.db import models, transaction
+from django.db import transaction
+from django.db.models.query import QuerySet
 
 from django_nitro_mailer.models import Email
+from django_nitro_mailer.settings import NITRO_EMAIL_BACKEND
 from django_nitro_mailer.utils import send_email_message, throttle_email_delivery
 
 logger = logging.getLogger(__name__)
 
 
-def send_emails(queryset: Optional[models.QuerySet] = None) -> None:
+@dataclass
+class SendEmailsResult:
+    success_count: int
+    failure_count: int
+
+def send_emails(queryset: QuerySet | None = None) -> SendEmailsResult:
     if queryset is None:
         queryset = Email.objects.exclude(priority=Email.Priorities.DEFERRED).order_by("-priority", "created_at")
 
-    connection = get_connection()
+    result = SendEmailsResult(success_count=0, failure_count=0)
+
+    connection = get_connection(NITRO_EMAIL_BACKEND)
     with transaction.atomic():
         for email_obj in queryset.select_for_update(nowait=True):
-            try:
-                email_message = email_obj.email
-                if email_obj.email:
-                    send_email_message(email_message, connection)
+            email_message = email_obj.email
+            if email_obj.email:
+                if send_email_message(email_message, connection):
                     email_obj.delete()
-                    throttle_email_delivery()
+                    result.success_count += 1
                 else:
-                    logger.error("No email message found for Email object: %s", email_obj.id, exc_info=True)
+                    result.failure_count += 1
+            else:
+                logger.error("Unable to deserialize email", extra={"email_id": email_obj.id})
+                result.failure_count += 1
 
-            except Exception as e:
-                logger.error(
-                    "Failed to send or delete email %s: %s",
-                    email_obj.id,
-                    str(e),
-                    extra={"email_id": email_obj.id},
-                    exc_info=True,
-                )
+            throttle_email_delivery()
+
+    return result
 
 
-def retry_deferred() -> None:
+def retry_deferred() -> SendEmailsResult:
     deferred_emails = Email.objects.filter(priority=Email.Priorities.DEFERRED)
-    if deferred_emails.exists():
-        send_emails(deferred_emails)
+    return send_emails(deferred_emails)
